Reorganize entry points
